+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * This helps kwcoco work better as a standalone package.
 * Union now works with videos
 * New formal schema with jsonschema validation support
+* Add `autobuild` kwarg to `CocoDset.subset`
 
 
 ### Changed